Changed port to ushort

+++ Mirror/Runtime/NetworkServer.cs
@@ -85,17 +85,17 @@ namespace Mirror
             RegisterHandler(MsgType.Ping, NetworkTime.OnServerPing);
         }
 
-        public static bool Listen(int serverPort, int maxConnections)
+        public static bool Listen(ushort serverPort, int maxConnections)
         {
             return InternalListen(null, serverPort, maxConnections);
         }
 
-        public static bool Listen(string ipAddress, int serverPort, int maxConnections)
+        public static bool Listen(string ipAddress, ushort serverPort, int maxConnections)
         {
             return InternalListen(ipAddress, serverPort, maxConnections);
         }
 
-        internal static bool InternalListen(string ipAddress, int serverPort, int maxConnections)
+        internal static bool InternalListen(string ipAddress, ushort serverPort, int maxConnections)
         {
             Initialize();
